cdc: limit old value cache with memory capacity
